[Gravel] Code to seperate data from codegen and dump into file.
THis code understand the flow inside ./cpu -> copy function- > and tries to dump data into file(both coefficient and grid_data).
THe grid data is in progress of of this but I can dump currently 2 files.
The grid data here is not really modified the way it should be, but next plan is to add some flags in ./cpu --flag which can read data from
the input file and can even dump into file.